Get ready for card comparison
A few things in this one:

* Change the name of the #value method to #human_value. This is so we
can still compare the 'raw' values later.
* Add a 'protected' method to read the 'raw' value. This means that the
raw value can't be read from outside the class (so you couldn't do
Card.new(1, :spades).value) but it can be read from inside the class,
and from other objects of the same class. We'll see why this is
necessary in the next step.